+++ domain/src/main/kotlin/com/kefir/repositories/AccountTypeRepository.kt
@@ -3,4 +3,4 @@ package com.kefir.repositories
 import com.kefir.entities.AccountType
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface AccountTypeRepository : JpaRepository<AccountType, Long>
+interface AccountTypeRepository : JpaRepository<AccountType, Int>
