+++ domain/src/main/kotlin/com/kefir/repositories/BankRepository.kt
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface BankRepository : JpaRepository<Bank, Long>
+interface BankRepository : JpaRepository<Bank, Int>
